refactor: [BONUS] change location function handle
